refactor(main page): #Y-12 Изменил структуру файлов. Добавил scss
