+++ packages/client/package.json
@@ -31,6 +31,8 @@
     "jest-environment-jsdom": "^29.0.1",
     "lefthook": "^1.3.9",
     "prettier": "^2.7.1",
+    "sass": "^1.69.5",
+    "sass-loader": "^13.3.2",
     "ts-jest": "^28.0.8",
     "typescript": "^4.8.2",
     "vite": "^3.0.7"
