+++ packages/client/src/assets/styles/common.scss
