+++ packages/client/src/assets/styles/index.scss
@@ -1,5 +1,5 @@
 @import url(../fonts/index.css);
-@import url(../styles/common.css);
+@import url(../styles/common.scss);
 
 body {
   -webkit-font-smoothing: antialiased;
