+++ packages/client/src/components/menu/index.ts
@@ -1,0 +1,3 @@
+import { Menu } from './Menu'
+
+export default Menu
