--- packages/client/src/components/menu/menu.module.css
@@ -1,29 +1,0 @@
-.subtitle {
-  font-size: 42px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: normal;
-}
-
-.menu {
-  margin: 155px 0 76px 0;
-}
-
-.links {
-  list-style: none;
-  padding: 0;
-  margin: 0;
-  display: flex;
-  flex-direction: column;
-}
-
-.link {
-  margin: 10px 0;
-  font-size: 24px;
-  text-decoration: none;
-  color: inherit;
-}
-
-.link:hover {
-  opacity: 0.8;
-}
