+++ packages/client/src/pages/mainPage/index.ts
@@ -1,0 +1,3 @@
+import { MainPage } from './MainPage'
+
+export default MainPage
